Move the method signature checks to initialization
This method is really looking at definition level stuff i.e whether the
job class has the necessary methods and the method declarations are
correct. As such, I think having this sort of check in the constructor
meakes more sense.
This also updates all the tests accordingly.

If this isn't the case/there's a case for not having it in the
constructor we can just use a boolean to not re-check the signatures on
every invocation of `perform`.