Optional `sorbet-runtime` support for interface validation
Methods that have Sorbet signatures are wrapped by `sorbet-runtime` so
that they can be type-checked for correct params/return-value at
runtime. However, that wrapper does not and cannot relay the whole
information about the original method like `arity` or `parameters`.
(Ref: https://github.com/sorbet/sorbet/issues/2643)

The workaround is to access the original method from the signature
if we detect Sorbet is activated and the method in question has a
signature.

This commit abstracts the method parameter extraction into a separate
method in order to carry out that workaround.