Fix CsvEnumerator #batches so it increments the cursor
I'm addressing an off-by-one-error in the iteration of the CsvEnumerator.

 When the job would resume after interruption,
it would begin by re-processing the record that was just sucessfully imported prior to interruption.
That is, the job would re-process the last iteration
instead of the moving onto the next iteration as expected.

This problem is happening because the CsvEnumerator passes the given cursor position
directly into #drop without incrementing it first.
not the index of the element we want to drop.
Since we just pass in the index of the last sucessful iteration and #drop isn't positional,
we end up removing all the elements before the last iteration,
but not the last iteration itself.

- I handle the cursor in the same way as the #build_array_enumerator.
  I increment the cursor by 1 before passing it to #drop,
  unless the cursor is nil, in which case 0 is passed instead