Bump version & update CHANGELOG.md
Only relevant change is enforcing serializable cursor. Non-serializable
cursors were always an "error", and their behaviour in the host job was
"undefined", depending on the object and how the host job handled it.
This change just makes the error loud and immediate, instead of silent
and delayed until resumption following interruption; hence the patch
version increment.