Fix YARD documentation error
This fixes the following error:

  [warn]: Load Order / Name Resolution Problem on Resque::Worker:
  -
  Something is trying to call mixins on object Resque::Worker before it has been recognized.
  This error usually means that you need to modify the order in which you parse files
  so that Resque::Worker is parsed before methods or other objects attempt to access it.
  -
  YARD will recover from this error and continue to parse but you *may* have problems
  with your generated documentation. You should probably fix this.
  -

  [error]: Unhandled exception in YARD::Handlers::Ruby::MixinHandler:
    in `lib/job-iteration/integrations/resque.rb`:14:

  	14: Resque::Worker.prepend(ResqueIterationExtension)

  [error]: ProxyMethodError: Proxy cannot call method #mixins on object 'Resque::Worker'

This is the approach documented in https://github.com/lsegal/yard/issues/1353#issuecomment-700188928

What's happening is that YARD thinks Resque::Worker is a constant of ours and
is erroring because it's being used before being defined (and documented).

By reopening the class instead, we ensure YARD knows about it, and by using the
"@private" magic comment, we ensure YARD doesn't include it in the
documentation it generates.

By addressing the error, we can stop ignoring the files.