Continue other jobs on Rails edge failures
Rails edge (and Ruby HEAD, but we always combine that with Rails edge)
is a moving target out of our control. Therefore, if a failure occurs on
a job testing against it, it is useful to let the other jobs continue,
as that helps narrow down if the failure is due to a change made there.